Created molecule for scroll link, hooked up within Hero
