+++ src/components/molecules/index.js
@@ -2,5 +2,6 @@ export ClientList from './ClientList';
 export DescriptiveList from './DescriptiveList';
 export GlobalNavigation from './GlobalNavigation';
 export PostExcerpt from './PostExcerpt';
+export ScrollLink from './ScrollLink';
 export ServiceList from './ServiceList';
 export TeamList from './TeamList';
